# openapi.yaml
openapi: 3.0.3
info:
  title: WEBonTour API
  description: API for the WEBonTour CMS
  version: 1.0.0
  contact:
    name: Project Team
  license:
    name: Project License
servers:
  - url: /api
    description: API base path
...
paths:
  /auth/signup:
    post:
      tags:
        - auth
      summary: Register a new user
      description: Create a new user account (protected by signup key or open registration)
      operationId: signupUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - email
              properties:
                username:
                  type: string
                  example: johndoe
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: password123
                email:
                  type: string
                  format: email
                  example: john@example.com
                signupKey:
                  type: string
                  description: Required for invite-only registration mode
                masterKey:
                  type: string
                  description: Master key for signup authorization
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                  message:
                    type: string
                    example: User registered successfully
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid signup key or unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /signup-keys:
    get:
      tags:
        - signup-keys
      summary: Get all signup keys
      description: Get a list of all signup keys (admin only)
      operationId: getSignupKeys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signup keys retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SignupKey'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - signup-keys
      summary: Create a new signup key
      description: Create a new signup key (admin only)
      operationId: createSignupKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  example: For John from HR
                  description: Optional note to track the key purpose
      responses:
        '201':
          description: Signup key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupKey'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /signup-keys/{id}:
    put:
      tags:
        - signup-keys
      summary: Update signup key note
      description: Update the note for a signup key (admin only)
      operationId: updateSignupKey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Signup key ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  example: Updated note
      responses:
        '200':
          description: Signup key updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupKey'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Signup key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - signup-keys
      summary: Delete signup key
      description: Delete a signup key (admin only)
      operationId: deleteSignupKey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Signup key ID
      responses:
        '200':
          description: Signup key deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Signup key deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Signup key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /signup-keys/validate:
    post:
      tags:
        - signup-keys
      summary: Validate signup key
      description: Validate if a signup key is valid (public endpoint)
      operationId: validateSignupKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - key
              properties:
                key:
                  type: string
                  example: abc123def456...
      responses:
        '200':
          description: Signup key validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  keyId:
                    type: integer
                    example: 5
        '400':
          description: Signup key is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invalid signup key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Site Settings endpoints
...
  /setup/status:
    get:
      tags:
        - setup
      summary: Check setup status
      description: Check if initial setup is required (no users exist)
      operationId: getSetupStatus
      responses:
        '200':
          description: Setup status retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  needsSetup:
                    type: boolean
                    example: true
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /setup/create-admin:
    post:
      tags:
        - setup
      summary: Create admin user
      description: Create the initial admin user using the configured master setup key (only works when no users exist)
      operationId: createAdminUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - email
                - masterKey
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: securepassword123
                email:
                  type: string
                  format: email
                  example: admin@example.com
                masterKey:
                  type: string
                  format: password
                  description: Value of the configured MASTER_SIGNUP_KEY
      responses:
        '201':
          description: Admin user created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                  message:
                    type: string
                    example: Admin user created successfully
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Bad request or setup already completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Invalid setup key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many setup attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: johndoe
        email:
          type: string
          format: email
          example: john@example.com
        role:
          type: string
          example: user
          enum: [user, writer, moderator, admin]
        emailNotifications:
          type: boolean
          description: Whether new-post emails are enabled
...
    SignupKey:
      type: object
      properties:
        id:
          type: integer
          example: 1
        key_value:
          type: string
          example: abc123def456ghi789...
        note:
          type: string
          example: For John from HR
          nullable: true
        created_at:
          type: string
          format: date-time
        created_by_username:
          type: string
          example: admin
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message